refactor: make Tank.owner_type required and fix set_position annotation

+++ src/core/game_object.py
@@ -62,7 +62,7 @@ class GameObject:
         """Get the current position of the object."""
         return (self.x, self.y)
 
-    def set_position(self, x: int, y: int) -> None:
+    def set_position(self, x: float, y: float) -> None:
         """
         Set the position of the object.
